Move run_program() etc. from proc_control.cpp back to util.cpp. This required changing a bunch of project files, and was turning into a rabbit hole.
