+++ client/app.h
@@ -38,6 +38,7 @@
 
 struct CLIENT_STATE;
 struct ASYNC_COPY;
+typedef int PROCESS_ID;
 
 #define MAX_STDERR_LEN  65536
     // The stderr output of an application is truncated to this length
@@ -61,7 +62,7 @@ struct ACTIVE_TASK {
     RESULT* result;
     WORKUNIT* wup;
     APP_VERSION* app_version;
-    int pid;
+    PROCESS_ID pid;
     PROCINFO procinfo;
 
     // START OF ITEMS SAVED IN TASK STATE FILE
