+++ client/makefile_sim
@@ -66,7 +66,6 @@ OBJS = \
     ../lib/opencl_boinc.o \
     ../lib/parse.o \
     ../lib/prefs.o \
-    ../lib/proc_control.o \
     ../lib/project_init.o \
     ../lib/proxy_info.o \
     ../lib/shmem.o \
