+++ clientgui/BOINCClientManager.cpp
@@ -24,7 +24,6 @@
 #include "diagnostics.h"
 #include "miofile.h"
 #include "str_replace.h"
-#include "proc_control.h"
 #include "util.h"
 
 #include "LogBOINC.h"
