+++ mac_build/HowToBuildBOINC_XCode.rtf
@@ -1,4 +1,4 @@
-{\rtf1\ansi\ansicpg1252\cocoartf2709
+{\rtf1\ansi\ansicpg1252\cocoartf2638
 \cocoascreenfonts1\cocoatextscaling0\cocoaplatform0{\fonttbl\f0\fswiss\fcharset0 Helvetica-Bold;\f1\fswiss\fcharset0 Helvetica;\f2\fmodern\fcharset0 Courier;
 \f3\fswiss\fcharset0 Arial-BoldMT;\f4\fnil\fcharset0 Menlo-Regular;\f5\fswiss\fcharset0 Helvetica-Oblique;
 \f6\fnil\fcharset0 LucidaGrande;}
@@ -8,7 +8,7 @@
 {\*\expandedcolortbl;;\cssrgb\c0\c38136\c99824;\csgenericrgb\c72941\c0\c0;\csgenericrgb\c100000\c100000\c100000;
 \csgenericrgb\c5490\c5490\c100000;\csgenericrgb\c96078\c96078\c96078;\cssrgb\c0\c0\c0;\csgray\c100000;\csgenericrgb\c96471\c96471\c96471;
 }
-\margl1440\margr1440\vieww14780\viewh11680\viewkind0
+\margl1440\margr1440\vieww14780\viewh11840\viewkind0
 \pard\tx720\tx1440\tx2160\tx2880\tx3600\tx4320\tx5040\tx5760\tx6480\tx7200\tx7920\tx8640\qc\partightenfactor0
 
 \f0\b\fs28 \cf0 Building BOINC Client and Manager on Macintosh OSX\
@@ -112,7 +112,7 @@ Note: building BOINC Manager 7.3.0 and later requires the OS 10.8 SDK or later.\
 \cf0 BOINC depends on seven third-party libraries: wxWidgets-3.1.5, c-ares-1.17.2, curl-7.79.1, openssl-3.0.0, freetype-2.11.0 and ftgl-2.1.3~rc5.  You can obtain the source files from the following URLs.  Clicking on the first URL of each pair will download the tar file.  The second URL will open the third party\'92s home web page.  On Mac OS X the tar file will usually be downloaded into the Downloads folder.  You will need to expand the tar files by double-clicking on them, which will create a folder and place the appropriate files into that folder. You will need to move these folders later.\
 \
 wxWidgets-3.1.5 (needed  only if you are building the BOINC Manager):\
-	{\field{\*\fldinst{HYPERLINK "https://github.com/wxWidgets/wxWidgets/releases/download/v3.1.6/wxWidgets-3.1.6.tar.bz2"}}{\fldrslt 
+	{\field{\*\fldinst{HYPERLINK "https://github.com/wxWidgets/wxWidgets/releases/download/v3.1.5/wxWidgets-3.1.5.tar.bz2"}}{\fldrslt 
 \f2\fs26 \cf2 https://github.com/wxWidgets/wxWidgets/releases/download/v3.1.5/wxWidgets-3.1.5.tar.bz2}}
 \f2\fs26 \
