+++ samples/wrappture/wrappture.cpp
@@ -45,7 +45,7 @@
 #include "parse.h"
 #include "str_util.h"
 #include "str_replace.h"
-#include "proc_control.h"
+#include "util.h"
 #include "error_numbers.h"
 
 #include "wrappture.h"
