+++ tests/unit-tests/lib/test_util.cpp
@@ -17,7 +17,6 @@
 
 #include "gtest/gtest.h"
 #include "util.h"
-#include "proc_control.h"
 #include <math.h>
 
 using namespace std;
